Moved name generation into a separate module
This is handy since multiple modules need it
